requirements.txt
pymongo
duckdb
pandas
mysql-connector-python
faker
pyairbyte
dbt-duckdb